services/backend/requirements.txt
fastapi == 0.108.0
tortoise-orm == 0.20.0
aerich == 0.7.1
asyncpg == 0.29.0
uvicorn == 0.25.0

# dev
pytest == 7.4.4
httpx==0.26.0
